초기 예제를 tutorial 폴더에 옮기고 쇼핑몰 애플리케이션을 설정을 위한 압축을 풀었습니다.
