+++ tutorial/shop/src/main/java/com/shop/ShopApplication.java
@@ -1,0 +1,13 @@
+package com.shop;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ShopApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ShopApplication.class, args);
+	}
+
+}
