+++ tutorial/shop/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+# 애플리케이션 포트 설정
+server.port=80
